+++ src/aac_datasets/datasets/functional/macs.py
@@ -17,7 +17,7 @@ from aac_datasets.datasets.functional.common import DatasetCard, LinkInfoHash
 from aac_datasets.utils.download import download_file, hash_file
 from aac_datasets.utils.globals import _get_root
 
-pylog = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 
 MACSSubset = Literal["full"]
@@ -85,7 +85,7 @@ def load_macs_dataset(
     captions_fname = MACS_FILES["captions"]["fname"]
     captions_fpath = osp.join(macs_dpath, captions_fname)
     if verbose >= 2:
-        pylog.debug(f"Reading captions file {captions_fname}...")
+        logger.debug(f"Reading captions file {captions_fname}...")
 
     with open(captions_fpath, "r") as file:
         caps_data = yaml.safe_load(file)
@@ -93,7 +93,7 @@ def load_macs_dataset(
     tau_meta_fname = "meta.csv"
     tau_meta_fpath = osp.join(tau_meta_dpath, tau_meta_fname)
     if verbose >= 2:
-        pylog.debug(f"Reading Tau Urban acoustic scene meta file {tau_meta_fname}...")
+        logger.debug(f"Reading Tau Urban acoustic scene meta file {tau_meta_fname}...")
 
     with open(tau_meta_fpath, "r") as file:
         reader = csv.DictReader(file, delimiter="\t")
@@ -102,7 +102,7 @@ def load_macs_dataset(
     competence_fname = MACS_FILES["annotators_competences"]["fname"]
     competence_fpath = osp.join(macs_dpath, competence_fname)
     if verbose >= 2:
-        pylog.debug(f"Reading file {competence_fname}...")
+        logger.debug(f"Reading file {competence_fname}...")
 
     with open(competence_fpath, "r") as file:
         reader = csv.DictReader(file, delimiter="\t")
@@ -155,7 +155,7 @@ def load_macs_dataset(
 
     if verbose >= 1:
         msg = f"Dataset {MACSCard.PRETTY_NAME} ({subset}) has been loaded. {len(next(iter(raw_data.values())))=})"
-        pylog.info(msg)
+        logger.info(msg)
 
     return raw_data, annotator_id_to_competence
 
@@ -207,7 +207,7 @@ def download_macs_dataset(
 
         if not osp.isfile(fpath) or force:
             if verbose >= 1:
-                pylog.info(f"Downloading captions file '{fname}'...")
+                logger.info(f"Downloading captions file '{fname}'...")
 
             url = file_info["url"]
             download_file(url, fpath, verbose=verbose)
@@ -221,7 +221,7 @@ def download_macs_dataset(
                     f"Please try to remove manually the file '{fpath}' and rerun {MACSCard.PRETTY_NAME} download."
                 )
             elif verbose >= 2:
-                pylog.debug(f"File '{fname}' has a valid checksum.")
+                logger.debug(f"File '{fname}' has a valid checksum.")
 
     captions_fpath = osp.join(macs_dpath, MACS_FILES["captions"]["fname"])
     with open(captions_fpath, "r") as file:
@@ -236,7 +236,7 @@ def download_macs_dataset(
         if not osp.isfile(zip_fpath) or force:
             if verbose >= 1:
                 msg = f"Downloading audio zip file '{zip_fpath}'... ({i+1}/{len(MACS_ARCHIVES_FILES)})"
-                pylog.info(msg)
+                logger.info(msg)
 
             url = file_info["url"]
             download_file(url, zip_fpath, verbose=verbose)
@@ -250,7 +250,7 @@ def download_macs_dataset(
                     f"Please try to remove manually the file '{zip_fpath}' and rerun {MACSCard.PRETTY_NAME} download."
                 )
             elif verbose >= 2:
-                pylog.debug(f"File '{zip_fname}' has a valid checksum.")
+                logger.debug(f"File '{zip_fname}' has a valid checksum.")
 
     # Extract files from TAU Urban Sound archives
     macs_fnames = dict.fromkeys(data["filename"] for data in captions_data)
@@ -260,7 +260,7 @@ def download_macs_dataset(
 
         if verbose >= 2:
             msg = f"Check to extract TAU Urban acoustic scenes archive {zip_fname=}..."
-            pylog.debug(msg)
+            logger.debug(msg)
 
         is_audio_archive = name.startswith("audio")
         if is_audio_archive:
@@ -281,7 +281,7 @@ def download_macs_dataset(
 
             if verbose >= 1:
                 msg = f"Extracting {len(members_to_extract)}/{len(file.namelist())} audio files from ZIP file '{zip_fname}'... ({i+1}/{len(MACS_ARCHIVES_FILES)})"
-                pylog.info(msg)
+                logger.info(msg)
 
             if len(members_to_extract) > 0:
                 file.extractall(archives_dpath, members_to_extract)
@@ -292,14 +292,14 @@ def download_macs_dataset(
 
     if clean_archives:
         if verbose >= 1:
-            pylog.info(f"Removing archives files in {archives_dpath}...")
+            logger.info(f"Removing archives files in {archives_dpath}...")
         shutil.rmtree(archives_dpath, ignore_errors=True)
 
     audio_fnames = [name for name in os.listdir(audio_dpath) if name.endswith(".wav")]
     assert len(audio_fnames) == len(macs_fnames)
 
     if verbose >= 2:
-        pylog.debug(f"Dataset {MACSCard.PRETTY_NAME} ({subset}) has been prepared.")
+        logger.debug(f"Dataset {MACSCard.PRETTY_NAME} ({subset}) has been prepared.")
 
 
 def download_macs_datasets(